+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Refactored & optimized HL provider (highlight should be faster now)
 - Override default keybindings only when neovim was initialized succesfully ( #112 )
 - Don't preselect `'<,'>` marks when invoking cmdline from visual line ( #111 )
+- Implemented commandline history ( #88 )
 
 ## [0.0.42]